requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
psycopg[binary]>=3.1,<4
psycopg-pool>=3.1,<4
alembic>=1.13,<2
pydantic>=2.0,<3
pydantic[email]>=2.0,<3
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
bcrypt>=4.0,<4.1
python-multipart>=0.0.9
openpyxl>=3.1,<4
pymupdf>=1.24.0
python-dotenv>=1.0,<2
slowapi>=0.1.8,<1